+++ packages/pres/test/presm.test.js
@@ -5,9 +5,10 @@ import { Presm }                 from '../src/Presm.js'
 
 test('presm test', () => {
   const conf = { str: SUBTLE, pos: OCEAN, neg: PAGODA }
-  const presm = Presm.build(SUBTLE, OCEAN, PAGODA)
+  const presm = Presm.build(conf.str, conf.pos, conf.neg)
   console.log(conf)
   console.log(presm)
+  console.log(presm.toString())
   console.log(deco(presm))
 
   console.log($.xbd(presm.xbd).ybd(presm.ybd).zbd(presm.zbd))