+++ packages/pres/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @palett/presets
 
+## 0.12.17
+
+### Patch Changes
+
+- fix dependency issue in @palett/projector
+- Updated dependencies
+  - @palett/color-algebra@0.12.17
+  - @palett/convert@0.12.17
+  - @palett/stringify@0.12.17
+
 ## 0.12.16
 
 ### Patch Changes